New functionality:
* Add function S2ShapeIndex Dump function
* Add Init to Graph nested classes
* S2Polygon::InitToOperation: Propagate S2Error

Bug fixes:
* Fix S2BooleanOperation empty/full handling
* Minor compilation fixes for nacl
* Fix LaxPolygonLayer bug
* Fix rare bug in S2BooleanOperation
* Fix rare bug in VisitCrossingEdgePairs
* Fix UB discovered by -D_GLIBCXX_DEBUG
* S2Cap::Add: Fix bug with empty other cap
* OwningShape: Fix potential memory leak
* S2Builder: Clear layer_is_full_polygon_predicates_
* S2EdgeTesselator: Improve error estimation

+++ src/s2/s2point_span.h
@@ -49,7 +49,7 @@ class S2PointLoopSpan : public S2PointSpan {
   reference operator[](int i) const noexcept {
     S2_DCHECK_GE(i, 0);
     S2_DCHECK_LT(i, 2 * size());
-    int j = i - size();
+    int j = i - static_cast<int>(size());
     return S2PointSpan::operator[](j < 0 ? i : j);
   }
 };
